+++ formula-android/src/main/java/com/instacart/formula/android/FragmentFlowState.kt
@@ -1,10 +1,10 @@
 package com.instacart.formula.android
 
 /**
- * Represents currently [activeKeys] and their [states].
+ * Represents currently [activeIds] and their [states].
  *
- * @param activeKeys Fragment contracts that are running their state management.
- * @param visibleKeys Fragment contracts that are currently visible to the user.
+ * @param activeIds Fragment contracts that are running their state management.
+ * @param visibleIds Fragment contracts that are currently visible to the user.
  * @param states Last emitted state of each active [FragmentKey].
  */
 data class FragmentFlowState(
@@ -13,13 +13,5 @@ data class FragmentFlowState(
     val features: Map<FragmentId, FeatureEvent> = emptyMap(),
     val states: Map<FragmentId, FragmentState> = emptyMap()
 ) {
-    @Deprecated("use activeIds")
-    val activeKeys: List<FragmentKey>
-        get() = activeIds.map { it.key }
-
-    @Deprecated("use visibleIds")
-    val visibleKeys: List<FragmentKey>
-        get() = visibleIds.map { it.key }
-
     fun visibleState() = visibleIds.lastOrNull()?.let { states[it] }
 }